refactor(assertions): keep metric rendering in templates util
